test: split PDF citation annotation ownership cases
